fix(bbdo): replace index_id with uint64 type (#553)
* fix(conflict_manager_storage): replace index_data in uint64
* enh(status.hh): change type for index_data
* enh(storage:status): index_is is uint64 now.
* enh(bbdo/stream): add ulong in serialize and unserialize + unit test is ok
* enh(conflict_manager): replace index_id in uint64
* enh(conflict_manager): fix troncated index_id in metrics table
* enh(storage): fix syntax

REFS: MON-6801

+++ storage/inc/com/centreon/broker/storage/rebuilder.hh
@@ -56,7 +56,7 @@ class rebuilder {
 
   // Local types.
   struct index_info {
-    uint32_t index_id;
+    uint64_t index_id;
     uint32_t host_id;
     uint32_t service_id;
     uint32_t rrd_retention;
@@ -78,11 +78,11 @@ class rebuilder {
                        uint32_t interval,
                        unsigned length);
   void _rebuild_status(mysql& ms,
-                       uint32_t index_id,
+                       uint64_t index_id,
                        uint32_t interval,
                        uint32_t length);
   void _send_rebuild_event(bool end, uint32_t id, bool is_index);
-  void _set_index_rebuild(mysql& db, uint32_t index_id, short state);
+  void _set_index_rebuild(mysql& db, uint64_t index_id, short state);
   void _run(asio::error_code ec);
 
  public:
